+++ src/main/java/kitchenpos/application/OrderService.java
@@ -7,13 +7,13 @@ import kitchenpos.application.dto.response.OrderDto;
 import kitchenpos.application.dto.request.UpdateOrderStatusDto;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuHistory;
-import kitchenpos.domain.repository.MenuHistoryRepository;
-import kitchenpos.domain.repository.MenuRepository;
+import kitchenpos.domain.menu.repository.MenuHistoryRepository;
+import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
-import kitchenpos.domain.repository.OrderLineItemRepository;
-import kitchenpos.domain.repository.OrderRepository;
-import kitchenpos.domain.repository.OrderTableRepository;
+import kitchenpos.domain.order.repository.OrderLineItemRepository;
+import kitchenpos.domain.order.repository.OrderRepository;
+import kitchenpos.domain.table.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
