+++ src/main/java/kitchenpos/domain/table/repository/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.repository;
+package kitchenpos.domain.table.repository;
 
 import java.util.List;
 import java.util.Optional;
