+++ src/main/java/kitchenpos/domain/table/repository/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.repository;
+package kitchenpos.domain.table.repository;
 
 import java.util.List;
 import kitchenpos.domain.table.TableGroup;
