+++ src/test/java/kitchenpos/domain/repository/MenuHistoryRepositoryTest.java
@@ -6,6 +6,8 @@ import java.math.BigDecimal;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuHistory;
 import kitchenpos.domain.menu.Price;
+import kitchenpos.domain.menu.repository.MenuHistoryRepository;
+import kitchenpos.domain.menu.repository.MenuRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
